Create anxiety
This is a visual representation of how thoughts can infiltrate the mind with anxiety and sometimes uncontrolled enragement over the many thoughts that come from trying to focus at times. It showcases how chaotic it can be and how random apparent they become in your mind, that it becomes difficult to concentrate on returning back to a calm state of mind.